Added parallax subsurface effect to the solid block shader.
Currently used by blue/packed ice, diamond block/ore, emerald block/ore, and the active redstone lamp.
The deepslate ores still need to be done.

The effect requires the block to be mapped to the top left quadrant of a 64x64px texture, where the top left is the surface color(RGB), top right is the subsurface height (R), subsurface visibility (G), and subsurface emissive intensity (B), the bottom left is the subsurface color (RGB), and the bottom right is "misc settings" (R), and chromatic aberration amount (G). The "misc settings" should be either 0 (the subsurface texture repeats), or 1 (the UVs are clamped to the edges of the subsurface texture). The alpha channel is unused, except for in the top left quadrant, where the alpha must be 248 where the effect should be used. All parts of the texture can be animated.