Revert "Test GH Actions for zipping the pack."
This reverts commit 741df76cb5f6dbf2b680e749929ac550bcad8eef.

It seemed to work fine, except something happened with the download. It got stuck at 64 KB and the zip only contained a couple of the files in the root folder, even though the workflow page said it should be 48 MB. Getting rid of the action for now, but could add it back if a fix is found for that problem.